256-byte multibyte program fix

+++ STM32/Core/Src/dumper.c
@@ -156,7 +156,7 @@ void write_flash(uint16_t address, uint16_t len, uint8_t *data)
   led_red();
   while (len > 0)
   {
-    uint8_t count = 0;
+    uint16_t count = 0;
     uint8_t *d = data;
     uint16_t a = address;
     uint16_t last_address = 0;
